Moved all calls into traits; Added calls for address endpoints; Added related address DTOs; Added calls for tracking endpoints;
