+++ src/DTO/ParsedAddressEntity.php
@@ -1,0 +1,22 @@
+<?php
+
+namespace BluefynInternational\ShipEngine\DTO;
+
+use BluefynInternational\ShipEngine\DTO\Validators\GreaterThanOrEquals;
+use BluefynInternational\ShipEngine\DTO\Validators\InArray;
+use BluefynInternational\ShipEngine\DTO\Validators\MaxLength;
+use BluefynInternational\ShipEngine\DTO\Validators\MinLength;
+use Spatie\DataTransferObject\DataTransferObject;
+
+class ParsedAddressEntity extends DataTransferObject
+{
+    public string $type;
+    #[GreaterThanOrEquals(0)]
+    public float $score;
+    public string $text;
+    #[GreaterThanOrEquals(0)]
+    public int $start_index;
+    #[GreaterThanOrEquals(0)]
+    public int $end_index;
+    public array $result;
+}
